Added global variables
1. Added global variables (fixes #12)
Global variables are stored at the beginning of the tape (left-most cells)
Because functions might access them, had to change functions structure. They are no longer position-independent.
They need to know where on the tape they are running from, in order to know how far left they need to go to reach the global variables.
This affects the function's code and because of that, we are now compiling a function on every call to it.

2. Divided Compiler.py to several smaller modules (fixes #9)
Included description at the beginning of each one

3. Fixed a bug where function parameters evaluation was assuming it runs from an incorrect place on the stack.
It happened in case there was a function call as a right-side operation of an expression, that takes a variable as parameter (e.g 1+foo(x))

4. Added global_variables.code to the examples directory

5. Updated game tic tac toe to use global variables